tests: Migrate git workflow tests to Jest
