+++ test/jest-stdout-serializer.js
@@ -1,0 +1,11 @@
+// my-serializer-module
+module.exports = {
+    print(val, serialize, indent) {
+        return val.stdout
+    },
+
+    test(val) {
+        return val;
+    },
+
+};
